+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Package `winner` was moved to `wal-windows`.
+- Package `wal-line` is now installed using `package-vc-install`.
+  `junk` recipes also use that function.
 
 ### Fixed
 
@@ -30,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   `dictionary`.
 - Package `popper`.
 - Package `diminish`.
+- Packages `quelpa` and `quelpa-use-package`.
 
 ## [v1.11.5]